Remove player_primary.lua hook
- update readme as a result
- bump minor version

The purpose for this change is to be more compatible with other mods.  bk3k
pointed out that although my hook design is more comprehensive, it's not
compatible with many pre-existing mods who have tackled the same problem via a
different solution.  Thus wrapping vanilla files where they can be patched is
the best way to "hook" into them.